task-84719 Worked on the following points:
Allowed seller/admin to download the attahcments
Added checkes if attachments at catalog level then admin/seller can not delete the attachments
Displayed the attahcments and links on seller product (in seller/admin side),
hide the delete option, remove the fields from the upload form if attachements allowed with catalog products

Tested with different use cases. :(

+++ admin-application/views/seller-products/inventory-digital-download-attachments-list.php
@@ -25,8 +25,33 @@ foreach ($records as $sn => $row) {
             case 'listserial':
                 $td->appendElement('plaintext', array(), $sr_no, true);
                 break;
+            case 'mainfile':
+                $td->appendElement('plaintext', array(), $row[$key], true);
+                $td->appendElement(
+                    "a",
+                    array(
+                        'class' => 'btn btn-clean btn-sm btn-icon',
+                        'title' => Labels::getLabel('LBL_download', $adminLangId),
+                        'href' => UrlHelper::generateUrl('SellerProducts', 'downloadAttachment', array($row['afile_id'], $recordId, $downloadrefType, 0, $row['mainfile'])),
+                        'target' => '_blank'
+                    ),
+                    '<i class="fa fa-download  icon"></i>',
+                    true
+                );
+                break;
             case 'preview':
-                $td->appendElement('plaintext', array(), $row['preview'], true);
+                $td->appendElement('plaintext', array(), $row[$key], true);
+                $td->appendElement(
+                    "a",
+                    array(
+                        'class' => 'btn btn-clean btn-sm btn-icon',
+                        'title' => Labels::getLabel('LBL_download', $adminLangId),
+                        'href' => UrlHelper::generateUrl('SellerProducts', 'downloadAttachment', array($row['prev_afile_id'], $recordId, $downloadrefType, 1, $row['preview'])),
+                        'target' => '_blank'
+                    ),
+                    '<i class="fa fa-download  icon"></i>',
+                    true
+                );
                 break;
             case 'pddr_options_code':
                 if (array_key_exists($row['pddr_options_code'], $options)) {
@@ -44,16 +69,18 @@ foreach ($records as $sn => $row) {
                 $td->appendElement('plaintext', array(), $lang_name, true);
                 break;
             case 'action':
-                $td->appendElement(
-                    "a",
-                    array(
-                        'class' => 'btn btn-clean btn-sm btn-icon',
-                        'title' => Labels::getLabel('LBL_Delete', $adminLangId),
-                        'onclick' => 'deleteDigitalFile(' . $row['afile_id'] . ', ' . $row['afile_record_id'] . ')', 'href' => 'javascript:void(0);'
-                    ),
-                    '<i class="fa fa-trash  icon"></i>',
-                    true
-                );
+                if (true === $canDelete) {
+                    $td->appendElement(
+                        "a",
+                        array(
+                            'class' => 'btn btn-clean btn-sm btn-icon',
+                            'title' => Labels::getLabel('LBL_Delete', $adminLangId),
+                            'onclick' => 'deleteDigitalFile(' . $row['afile_id'] . ', ' . $row['afile_record_id'] . ')', 'href' => 'javascript:void(0);'
+                        ),
+                        '<i class="fa fa-trash  icon"></i>',
+                        true
+                    );
+                }
                 if (empty($row['preview'])) {
                     $td->appendElement(
                         "a",
