+++ application/views/seller/inventory-digital-download-attachments-list.php
@@ -13,7 +13,7 @@ $th = $tbl->appendElement('thead')->appendElement('tr', array('class' => 'hide--
 foreach ($arr_flds as $val) {
     $e = $th->appendElement('th', array(), $val);
 }
-// CommonHelper::printArray([['file' => __FILE__, 'line' => __LINE__], $records], 1);
+
 $sr_no = 0;
 foreach ($records as $sn => $row) {
     $sr_no++;
@@ -32,8 +32,7 @@ foreach ($records as $sn => $row) {
                     array(
                         'class' => 'btn btn-clean btn-sm btn-icon',
                         'title' => Labels::getLabel('LBL_download', $siteLangId),
-                        // 'onclick' => 'downloadAttachment(' . $row['afile_id'] . ', ' . $selProdId . '); return false;',
-                        'href' => UrlHelper::generateUrl('Seller', 'downloadInventoryAttachment', array($row['afile_id'], $selProdId, 0, $row['mainfile'])),
+                        'href' => UrlHelper::generateUrl('Seller', 'downloadAttachment', array($row['afile_id'], $selProdId, Product::CATALOG_TYPE_INVENTORY, 0, $row['mainfile'])),
                         'target' => '_blank'
                     ),
                     '<i class="fa fa-download  icon"></i>',
@@ -47,8 +46,7 @@ foreach ($records as $sn => $row) {
                     array(
                         'class' => 'btn btn-clean btn-sm btn-icon',
                         'title' => Labels::getLabel('LBL_download', $siteLangId),
-                        /* 'onclick' => 'downloadAttachment(' . $row['prev_afile_id'] . ', ' . $selProdId . ', 1)', 'href' => 'javascript:void(0);', */
-                        'href' => UrlHelper::generateUrl('Seller', 'downloadInventoryAttachment', array($row['prev_afile_id'], $selProdId, 1, $row['preview'])),
+                        'href' => UrlHelper::generateUrl('Seller', 'downloadAttachment', array($row['prev_afile_id'], $selProdId, Product::CATALOG_TYPE_INVENTORY, 1, $row['preview'])),
                         'target' => '_blank'
                     ),
                     '<i class="fa fa-download  icon"></i>',
@@ -71,16 +69,18 @@ foreach ($records as $sn => $row) {
                 $td->appendElement('plaintext', array(), $lang_name, true);
                 break;
             case 'action':
-                $td->appendElement(
-                    "a",
-                    array(
-                        'class' => 'btn btn-clean btn-sm btn-icon',
-                        'title' => Labels::getLabel('LBL_Delete', $siteLangId),
-                        'onclick' => 'deleteDigitalFile(' . $row['afile_id'] . ', ' . $row['afile_record_id'] . ')', 'href' => 'javascript:void(0);'
-                    ),
-                    '<i class="fa fa-trash  icon"></i>',
-                    true
-                );
+                if (true === $canDelete) {
+                    $td->appendElement(
+                        "a",
+                        array(
+                            'class' => 'btn btn-clean btn-sm btn-icon',
+                            'title' => Labels::getLabel('LBL_Delete', $siteLangId),
+                            'onclick' => 'deleteDigitalFile(' . $row['afile_id'] . ', ' . $row['afile_record_id'] . ')', 'href' => 'javascript:void(0);'
+                        ),
+                        '<i class="fa fa-trash  icon"></i>',
+                        true
+                    );
+                }
                 if (empty($row['preview'])) {
                     $td->appendElement(
                         "a",
